server/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
elevenlabs
boto3
pydantic
paramiko 